requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
PyJWT==2.10.1
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
